Add a /health handler which returns OK if the wallet and sumokoind deamons are reachable.
This handler makes it more transparent to miners whether a pool is in
fact healthy: It is possible today that a pool server gives work to
miners without having a wallet RPC daemon running in the background.
In this case the miners will go unpaid until the admin fixes this.

This handler allows miner software, and also sites like sumopools.com,
to check the health of the pool.