Implement ryo-address-validator tool (#2)
Implemented `ryo-address-validator` tool that ships with ryo into the pool software to validate miner's public addresses instead of relying on the prefix. This feature is optional, and is disabled by default, but when activated calls the tool for address checking. It will automatically determine testnet/mainnet addresses and if the user is mining to the wrong network.